requirements.txt
huggingface_hub
python-dotenv
requests
pandas
psycopg2-binary
